only try to decrypt if downloading return 200 OK

+++ hls-download.py
@@ -138,38 +138,39 @@ def main(argv):
                     d.perform()
                     d.close()
 
-                    # decrypt the file if necessary
-                    if segment.encryption_method == "AES-128":
-                        command = ["openssl", "aes-128-cbc", "-d",
-                                "-K", key_cache.get(segment.key_url),
-                                "-iv", "%032x" % segment.sequence_id,
-                                "-in", encrypted_segment_filename,
-                                "-out", segment_filename]
-                        printlog("decryption start")
-                        openssl_return_code = subprocess.call(command)
-
-                    # check decryption result
-                    is_decryption_successful = False
-                    if segment.encryption_method != None:
-                        if openssl_return_code == 0:
-                            with open(segment_filename, "rb") as decrypted_file:
-                                first_byte = decrypted_file.read(1)
-                                decrypted_file.close()
-                                is_decryption_successful = (first_byte == 'G') # the first byte should be 'G' (0x47)
-                            if is_decryption_successful:
-                                printlog("decryption finished")
-                                os.remove(encrypted_segment_filename)
+                    if d.response_status == 200:
+                        # decrypt the file if necessary
+                        if segment.encryption_method == "AES-128":
+                            command = ["openssl", "aes-128-cbc", "-d",
+                                    "-K", key_cache.get(segment.key_url),
+                                    "-iv", "%032x" % segment.sequence_id,
+                                    "-in", encrypted_segment_filename,
+                                    "-out", segment_filename]
+                            printlog("decryption start")
+                            openssl_return_code = subprocess.call(command)
+    
+                        # check decryption result
+                        is_decryption_successful = False
+                        if segment.encryption_method != None:
+                            if openssl_return_code == 0:
+                                with open(segment_filename, "rb") as decrypted_file:
+                                    first_byte = decrypted_file.read(1)
+                                    decrypted_file.close()
+                                    is_decryption_successful = (first_byte == 'G') # the first byte should be 'G' (0x47)
+                                if is_decryption_successful:
+                                    printlog("decryption finished")
+                                    os.remove(encrypted_segment_filename)
+                                else:
+                                    printlog("decryption failed, first byte of file is: 0x%x (expected to be 0x47)" % first_byte)
                             else:
-                                printlog("decryption failed, first byte of file is: 0x%x (expected to be 0x47)" % first_byte)
-                        else:
-                            printlog("Decryption failed, openssl returned: %d" % openssl_return_code)
-                            
-                    # print to logs if it is plaintext or decryption was successful
-                    if segment.encryption_method == None or is_decryption_successful:
-                        segment.is_download_successful = True
-                        with open("%s/%s-list.txt" % (data_dir, name), "a+") as list_file:
-                            list_file.write("%d,%d,%d,%s\n" % (segment.sequence_id, segment.timestamp, segment.duration, "%s-%d.ts" % (name, segment.sequence_id)))
-                            list_file.close()
+                                printlog("Decryption failed, openssl returned: %d" % openssl_return_code)
+                                
+                        # print to logs if it is plaintext or decryption was successful
+                        if segment.encryption_method == None or is_decryption_successful:
+                            segment.is_download_successful = True
+                            with open("%s/%s-list.txt" % (data_dir, name), "a+") as list_file:
+                                list_file.write("%d,%d,%d,%s\n" % (segment.sequence_id, segment.timestamp, segment.duration, "%s-%d.ts" % (name, segment.sequence_id)))
+                                list_file.close()
 
                 # reording duration
                 recorded_duration = recorded_duration + segment.duration
